API

Vulkan.AabbPositionsKHR

Vulkan.AccelerationStructureBuildOffsetInfoKHR

Vulkan.AccelerationStructureGeometryDataKHR

Vulkan.AccelerationStructureInstanceKHR

Vulkan.AttachmentDescription

Vulkan.AttachmentReference

Vulkan.AttachmentSampleLocationsEXT

Vulkan.BindIndexBufferIndirectCommandNV

Vulkan.BindShaderGroupIndirectCommandNV

Vulkan.BindVertexBufferIndirectCommandNV

Vulkan.BufferCopy

Vulkan.BufferImageCopy

Vulkan.ClearAttachment

Vulkan.ClearColorValue

Vulkan.ClearDepthStencilValue

Vulkan.ClearRect

Vulkan.ClearValue

Vulkan.CoarseSampleLocationNV

Vulkan.ComponentMapping

Vulkan.ConformanceVersion

Vulkan.DebugUtilsMessengerEXT

Vulkan.DescriptorBufferInfo

Vulkan.DescriptorImageInfo

Vulkan.DescriptorPoolSize

Vulkan.DescriptorUpdateTemplateEntry

Vulkan.DeviceOrHostAddressConstKHR

Vulkan.DeviceOrHostAddressKHR

Vulkan.DispatchIndirectCommand

Vulkan.DisplayModeParametersKHR

Vulkan.DrawIndexedIndirectCommand

Vulkan.DrawIndirectCommand

Vulkan.DrawMeshTasksIndirectCommandNV

Vulkan.Extent2D

Vulkan.Extent3D

Vulkan.GeometryDataNV

Vulkan.ImageBlit

Vulkan.ImageCopy

Vulkan.ImageResolve

Vulkan.ImageSubresource

Vulkan.ImageSubresourceLayers

Vulkan.ImageSubresourceRange

Vulkan.IndirectCommandsStreamNV

Vulkan.InputAttachmentAspectReference

Vulkan.Offset2D

Vulkan.Offset3D

Vulkan.PerformanceCounterResultKHR

Vulkan.PerformanceValueDataINTEL

Vulkan.PerformanceValueINTEL

Vulkan.PhysicalDeviceFeatures

Vulkan.PipelineColorBlendAttachmentState

Vulkan.PresentTimeGOOGLE

Vulkan.PushConstantRange

Vulkan.Rect2D

Vulkan.RectLayerKHR

Vulkan.SampleLocationEXT

Vulkan.SetStateFlagsIndirectCommandNV

Vulkan.Setup

Vulkan.SparseImageMemoryBind

Vulkan.SparseMemoryBind

Vulkan.SpecializationMapEntry

Vulkan.StencilOpState

Vulkan.StridedBufferRegionKHR

Vulkan.SubpassDependency

Vulkan.SubpassSampleLocationsEXT

Vulkan.TraceRaysIndirectCommandKHR

Vulkan.TransformMatrixKHR

Vulkan.VertexInputAttributeDescription

Vulkan.VertexInputBindingDescription

Vulkan.VertexInputBindingDivisorDescriptionEXT

Vulkan.Viewport

Vulkan.ViewportSwizzleNV

Vulkan.ViewportWScalingNV

Vulkan.VulkanApplication

Vulkan.XYColorEXT

Vulkan.acquire_full_screen_exclusive_mode_ext

Vulkan.acquire_next_image_2_khr

Vulkan.acquire_next_image_khr

Vulkan.acquire_performance_configuration_intel

Vulkan.acquire_profiling_lock_khr

Vulkan.acquire_xlib_display_ext

Vulkan.allocate_command_buffers

Vulkan.allocate_descriptor_sets

Vulkan.allocate_memory

Vulkan.begin_command_buffer

Vulkan.bind_acceleration_structure_memory_khr

Vulkan.bind_buffer_memory

Vulkan.bind_buffer_memory_2

Vulkan.bind_image_memory

Vulkan.bind_image_memory_2

Vulkan.build_acceleration_structure_khr

Vulkan.cmd_begin_conditional_rendering_ext

Vulkan.cmd_begin_debug_utils_label_ext

Vulkan.cmd_begin_query

Vulkan.cmd_begin_query_indexed_ext

Vulkan.cmd_begin_render_pass

Vulkan.cmd_begin_render_pass_2

Vulkan.cmd_begin_transform_feedback_ext

Vulkan.cmd_bind_descriptor_sets

Vulkan.cmd_bind_index_buffer

Vulkan.cmd_bind_pipeline

Vulkan.cmd_bind_pipeline_shader_group_nv

Vulkan.cmd_bind_shading_rate_image_nv

Vulkan.cmd_bind_transform_feedback_buffers_ext

Vulkan.cmd_bind_vertex_buffers

Vulkan.cmd_bind_vertex_buffers_2_ext

Vulkan.cmd_blit_image

Vulkan.cmd_build_acceleration_structure_indirect_khr

Vulkan.cmd_build_acceleration_structure_khr

Vulkan.cmd_build_acceleration_structure_nv

Vulkan.cmd_clear_attachments

Vulkan.cmd_clear_color_image

Vulkan.cmd_clear_depth_stencil_image

Vulkan.cmd_copy_acceleration_structure_khr

Vulkan.cmd_copy_acceleration_structure_nv

Vulkan.cmd_copy_acceleration_structure_to_memory_khr

Vulkan.cmd_copy_buffer

Vulkan.cmd_copy_buffer_to_image

Vulkan.cmd_copy_image

Vulkan.cmd_copy_image_to_buffer

Vulkan.cmd_copy_memory_to_acceleration_structure_khr

Vulkan.cmd_copy_query_pool_results

Vulkan.cmd_debug_marker_begin_ext

Vulkan.cmd_debug_marker_end_ext

Vulkan.cmd_debug_marker_insert_ext

Vulkan.cmd_dispatch

Vulkan.cmd_dispatch_base

Vulkan.cmd_dispatch_indirect

Vulkan.cmd_draw

Vulkan.cmd_draw_indexed

Vulkan.cmd_draw_indexed_indirect

Vulkan.cmd_draw_indexed_indirect_count

Vulkan.cmd_draw_indirect

Vulkan.cmd_draw_indirect_byte_count_ext

Vulkan.cmd_draw_indirect_count

Vulkan.cmd_draw_mesh_tasks_indirect_count_nv

Vulkan.cmd_draw_mesh_tasks_indirect_nv

Vulkan.cmd_draw_mesh_tasks_nv

Vulkan.cmd_end_conditional_rendering_ext

Vulkan.cmd_end_debug_utils_label_ext

Vulkan.cmd_end_query

Vulkan.cmd_end_query_indexed_ext

Vulkan.cmd_end_render_pass

Vulkan.cmd_end_render_pass_2

Vulkan.cmd_end_transform_feedback_ext

Vulkan.cmd_execute_commands

Vulkan.cmd_execute_generated_commands_nv

Vulkan.cmd_fill_buffer

Vulkan.cmd_insert_debug_utils_label_ext

Vulkan.cmd_next_subpass

Vulkan.cmd_next_subpass_2

Vulkan.cmd_pipeline_barrier

Vulkan.cmd_preprocess_generated_commands_nv

Vulkan.cmd_push_constants

Vulkan.cmd_push_descriptor_set_khr

Vulkan.cmd_push_descriptor_set_with_template_khr

Vulkan.cmd_reset_event

Vulkan.cmd_reset_query_pool

Vulkan.cmd_resolve_image

Vulkan.cmd_set_blend_constants

Vulkan.cmd_set_checkpoint_nv

Vulkan.cmd_set_coarse_sample_order_nv

Vulkan.cmd_set_cull_mode_ext

Vulkan.cmd_set_depth_bias

Vulkan.cmd_set_depth_bounds

Vulkan.cmd_set_depth_bounds_test_enable_ext

Vulkan.cmd_set_depth_compare_op_ext

Vulkan.cmd_set_depth_test_enable_ext

Vulkan.cmd_set_depth_write_enable_ext

Vulkan.cmd_set_device_mask

Vulkan.cmd_set_discard_rectangle_ext

Vulkan.cmd_set_event

Vulkan.cmd_set_exclusive_scissor_nv

Vulkan.cmd_set_front_face_ext

Vulkan.cmd_set_line_stipple_ext

Vulkan.cmd_set_line_width

Vulkan.cmd_set_performance_marker_intel

Vulkan.cmd_set_performance_override_intel

Vulkan.cmd_set_performance_stream_marker_intel

Vulkan.cmd_set_primitive_topology_ext

Vulkan.cmd_set_sample_locations_ext

Vulkan.cmd_set_scissor

Vulkan.cmd_set_scissor_with_count_ext

Vulkan.cmd_set_stencil_compare_mask

Vulkan.cmd_set_stencil_op_ext

Vulkan.cmd_set_stencil_reference

Vulkan.cmd_set_stencil_test_enable_ext

Vulkan.cmd_set_stencil_write_mask

Vulkan.cmd_set_viewport

Vulkan.cmd_set_viewport_shading_rate_palette_nv

Vulkan.cmd_set_viewport_w_scaling_nv

Vulkan.cmd_set_viewport_with_count_ext

Vulkan.cmd_trace_rays_indirect_khr

Vulkan.cmd_trace_rays_khr

Vulkan.cmd_trace_rays_nv

Vulkan.cmd_update_buffer

Vulkan.cmd_wait_events

Vulkan.cmd_write_acceleration_structures_properties_khr

Vulkan.cmd_write_buffer_marker_amd

Vulkan.cmd_write_timestamp

Vulkan.compile_deferred_nv

Vulkan.copy_acceleration_structure_khr

Vulkan.copy_acceleration_structure_to_memory_khr

Vulkan.copy_memory_to_acceleration_structure_khr

Vulkan.debug_marker_set_object_name_ext

Vulkan.debug_marker_set_object_tag_ext

Vulkan.debug_report_message_ext

Vulkan.deferred_operation_join_khr

Vulkan.device_wait_idle

Vulkan.display_power_control_ext

Vulkan.end_command_buffer

Vulkan.enumerate_instance_version

Vulkan.flush_mapped_memory_ranges

Vulkan.free_command_buffers

Vulkan.free_descriptor_sets

Vulkan.free_memory

Vulkan.get_acceleration_structure_device_address_khr

Vulkan.get_acceleration_structure_handle_nv

Vulkan.get_acceleration_structure_memory_requirements_khr

Vulkan.get_acceleration_structure_memory_requirements_nv

Vulkan.get_android_hardware_buffer_properties_android

Vulkan.get_buffer_device_address

Vulkan.get_buffer_memory_requirements

Vulkan.get_buffer_memory_requirements_2

Vulkan.get_buffer_opaque_capture_address

Vulkan.get_calibrated_timestamps_ext

Vulkan.get_deferred_operation_max_concurrency_khr

Vulkan.get_deferred_operation_result_khr

Vulkan.get_descriptor_set_layout_support

Vulkan.get_device_acceleration_structure_compatibility_khr

Vulkan.get_device_group_peer_memory_features

Vulkan.get_device_group_present_capabilities_khr

Vulkan.get_device_group_surface_present_modes_2_ext

Vulkan.get_device_group_surface_present_modes_khr

Vulkan.get_device_memory_commitment

Vulkan.get_device_memory_opaque_capture_address

Vulkan.get_device_proc_addr

Vulkan.get_device_queue

Vulkan.get_device_queue_2

Vulkan.get_display_plane_capabilities_2_khr

Vulkan.get_display_plane_capabilities_khr

Vulkan.get_event_status

Vulkan.get_fence_fd_khr

Vulkan.get_fence_status

Vulkan.get_fence_win_32_handle_khr

Vulkan.get_generated_commands_memory_requirements_nv

Vulkan.get_image_drm_format_modifier_properties_ext

Vulkan.get_image_memory_requirements

Vulkan.get_image_memory_requirements_2

Vulkan.get_image_subresource_layout

Vulkan.get_image_view_address_nvx

Vulkan.get_image_view_handle_nvx

Vulkan.get_instance_proc_addr

Vulkan.get_memory_android_hardware_buffer_android

Vulkan.get_memory_fd_khr

Vulkan.get_memory_fd_properties_khr

Vulkan.get_memory_host_pointer_properties_ext

Vulkan.get_memory_win_32_handle_khr

Vulkan.get_memory_win_32_handle_nv

Vulkan.get_memory_win_32_handle_properties_khr

Vulkan.get_performance_parameter_intel

Vulkan.get_physical_device_external_buffer_properties

Vulkan.get_physical_device_external_fence_properties

Vulkan.get_physical_device_external_image_format_properties_nv

Vulkan.get_physical_device_external_semaphore_properties

Vulkan.get_physical_device_features

Vulkan.get_physical_device_features_2

Vulkan.get_physical_device_format_properties

Vulkan.get_physical_device_format_properties_2

Vulkan.get_physical_device_image_format_properties

Vulkan.get_physical_device_image_format_properties_2

Vulkan.get_physical_device_memory_properties

Vulkan.get_physical_device_memory_properties_2

Vulkan.get_physical_device_multisample_properties_ext

Vulkan.get_physical_device_properties

Vulkan.get_physical_device_properties_2

Vulkan.get_physical_device_queue_family_performance_query_passes_khr

Vulkan.get_physical_device_surface_capabilities_2_ext

Vulkan.get_physical_device_surface_capabilities_2_khr

Vulkan.get_physical_device_surface_capabilities_khr

Vulkan.get_physical_device_surface_support_khr

Vulkan.get_physical_device_wayland_presentation_support_khr

Vulkan.get_physical_device_win_32_presentation_support_khr

Vulkan.get_physical_device_xcb_presentation_support_khr

Vulkan.get_physical_device_xlib_presentation_support_khr

Vulkan.get_private_data_ext

Vulkan.get_query_pool_results

Vulkan.get_rand_r_output_display_ext

Vulkan.get_ray_tracing_capture_replay_shader_group_handles_khr

Vulkan.get_ray_tracing_shader_group_handles_khr

Vulkan.get_refresh_cycle_duration_google

Vulkan.get_render_area_granularity

Vulkan.get_semaphore_counter_value

Vulkan.get_semaphore_fd_khr

Vulkan.get_semaphore_win_32_handle_khr

Vulkan.get_swapchain_counter_ext

Vulkan.get_swapchain_status_khr

Vulkan.import_fence_fd_khr

Vulkan.import_fence_win_32_handle_khr

Vulkan.import_semaphore_fd_khr

Vulkan.import_semaphore_win_32_handle_khr

Vulkan.initialize_performance_api_intel

Vulkan.invalidate_mapped_memory_ranges

Vulkan.map_memory

Vulkan.merge_pipeline_caches

Vulkan.merge_validation_caches_ext

Vulkan.queue_begin_debug_utils_label_ext

Vulkan.queue_bind_sparse

Vulkan.queue_end_debug_utils_label_ext

Vulkan.queue_insert_debug_utils_label_ext

Vulkan.queue_present_khr

Vulkan.queue_set_performance_configuration_intel

Vulkan.queue_submit

Vulkan.queue_wait_idle

Vulkan.register_device_event_ext

Vulkan.register_display_event_ext

Vulkan.release_display_ext

Vulkan.release_full_screen_exclusive_mode_ext

Vulkan.release_performance_configuration_intel

Vulkan.release_profiling_lock_khr

Vulkan.reset_command_buffer

Vulkan.reset_command_pool

Vulkan.reset_descriptor_pool

Vulkan.reset_event

Vulkan.reset_fences

Vulkan.reset_query_pool

Vulkan.set_debug_utils_object_name_ext

Vulkan.set_debug_utils_object_tag_ext

Vulkan.set_event

Vulkan.set_hdr_metadata_ext

Vulkan.set_local_dimming_amd

Vulkan.set_private_data_ext

Vulkan.setup_viewport!

Vulkan.setup_viewport!

Vulkan.signal_semaphore

Vulkan.submit_debug_utils_message_ext

Vulkan.trim_command_pool

Vulkan.uninitialize_performance_api_intel

Vulkan.unmap_memory

Vulkan.update_descriptor_set_with_template

Vulkan.update_descriptor_sets

Vulkan.wait_for_fences

Vulkan.wait_semaphores

Vulkan.write_acceleration_structures_properties_khr

Vulkan.@check

Vulkan.AabbPositionsKHR — Method

Generic constructor.

Vulkan.AccelerationStructureBuildOffsetInfoKHR — Method

Generic constructor.

Vulkan.AccelerationStructureGeometryDataKHR — Method

Generic constructor.

Vulkan.AccelerationStructureInstanceKHR — Method

Generic constructor.

Vulkan.AttachmentDescription — Method

Generic constructor.

Vulkan.AttachmentReference — Method

Generic constructor.

Vulkan.AttachmentSampleLocationsEXT — Method

Generic constructor.

Vulkan.BindIndexBufferIndirectCommandNV — Method

Generic constructor.

Vulkan.BindShaderGroupIndirectCommandNV — Method

Generic constructor.

Vulkan.BindVertexBufferIndirectCommandNV — Method

Generic constructor.

Vulkan.BufferCopy — Method

Generic constructor.

Vulkan.BufferImageCopy — Method

Generic constructor.

Vulkan.ClearAttachment — Method

Generic constructor.

Vulkan.ClearColorValue — Method

Generic constructor.

Vulkan.ClearDepthStencilValue — Method

Generic constructor.

Vulkan.ClearRect — Method

Generic constructor.

Vulkan.ClearValue — Method

Generic constructor.

Vulkan.CoarseSampleLocationNV — Method

Generic constructor.

Vulkan.ComponentMapping — Method

Generic constructor.

Vulkan.ConformanceVersion — Method

Generic constructor.

Vulkan.DebugUtilsMessengerEXT — Method

Register a user callback and return the corresponding messenger.

A default named default_debug_callback can be converted to a function pointer to use as a callback.

Warning

callback must be a function pointer of type Ptr{Nothing} obtained from a callback_f function as follows: callback = @cfunction(callback_f, UInt32, (DebugUtilsMessageSeverityFlagBitsEXT, DebugUtilsMessageTypeFlagBitsEXT, Ptr{VkDebugUtilsMessengerCallbackDataEXT}, Ptr{Cvoid})) with callback_f a Julia function with the signature matching the @cfunction call.

Vulkan.DescriptorBufferInfo — Method

Generic constructor.

Vulkan.DescriptorImageInfo — Method

Generic constructor.

Vulkan.DescriptorPoolSize — Method

Generic constructor.

Vulkan.DescriptorUpdateTemplateEntry — Method

Generic constructor.

Vulkan.DeviceOrHostAddressConstKHR — Method

Generic constructor.

Vulkan.DeviceOrHostAddressKHR — Method

Generic constructor.

Vulkan.DispatchIndirectCommand — Method

Generic constructor.

Vulkan.DisplayModeParametersKHR — Method

Generic constructor.

Vulkan.DrawIndexedIndirectCommand — Method

Generic constructor.

Vulkan.DrawIndirectCommand — Method

Generic constructor.

Vulkan.DrawMeshTasksIndirectCommandNV — Method

Generic constructor.

Vulkan.Extent2D — Method

Generic constructor.

Vulkan.Extent3D — Method

Generic constructor.

Vulkan.GeometryDataNV — Method

Generic constructor.

Vulkan.ImageBlit — Method

Generic constructor.

Vulkan.ImageCopy — Method

Generic constructor.

Vulkan.ImageResolve — Method

Generic constructor.

Vulkan.ImageSubresource — Method

Generic constructor.

Vulkan.ImageSubresourceLayers — Method

Generic constructor.

Vulkan.ImageSubresourceRange — Method

Generic constructor.

Vulkan.IndirectCommandsStreamNV — Method

Generic constructor.

Vulkan.InputAttachmentAspectReference — Method

Generic constructor.

Vulkan.Offset2D — Method

Generic constructor.

Vulkan.Offset3D — Method

Generic constructor.

Vulkan.PerformanceCounterResultKHR — Method

Generic constructor.

Vulkan.PerformanceValueDataINTEL — Method

Generic constructor.

Vulkan.PerformanceValueINTEL — Method

Generic constructor.

Vulkan.PhysicalDeviceFeatures — Method

Generic constructor.

Vulkan.PipelineColorBlendAttachmentState — Method

Generic constructor.

Vulkan.PresentTimeGOOGLE — Method

Generic constructor.

Vulkan.PushConstantRange — Method

Generic constructor.

Vulkan.Rect2D — Method

Generic constructor.

Vulkan.RectLayerKHR — Method

Generic constructor.

Vulkan.SampleLocationEXT — Method

Generic constructor.

Vulkan.SetStateFlagsIndirectCommandNV — Method

Generic constructor.

Vulkan.Setup — Type

While handles are primordial for executing Vulkan API calls, they usually involve a lot more information that is intrinsically linked with the handle itself. A setup wraps a Handle by gathering the information that is required for a given piece of the application to work. For a logical device, that would include the physical device and the device itself (as a Handle), as well as its queues, semaphores and fences. A swapchain is also bound to a Handle, covers a certain extent of a surface and possesses images. When performing API calls, since a Setup object is bound to a handle, there is no need to specify the handle manually. The Setup object will automatically be converted to its handle when such call is performed.

Vulkan.SparseImageMemoryBind — Method

Generic constructor.

Vulkan.SparseMemoryBind — Method

Generic constructor.

Vulkan.SpecializationMapEntry — Method

Generic constructor.

Vulkan.StencilOpState — Method

Generic constructor.

Vulkan.StridedBufferRegionKHR — Method

Generic constructor.

Vulkan.SubpassDependency — Method

Generic constructor.

Vulkan.SubpassSampleLocationsEXT — Method

Generic constructor.

Vulkan.TraceRaysIndirectCommandKHR — Method

Generic constructor.

Vulkan.TransformMatrixKHR — Method

Generic constructor.

Vulkan.VertexInputAttributeDescription — Method

Generic constructor.

Vulkan.VertexInputBindingDescription — Method

Generic constructor.

Vulkan.VertexInputBindingDivisorDescriptionEXT — Method

Generic constructor.

Vulkan.Viewport — Method

Generic constructor.

Vulkan.ViewportSwizzleNV — Method

Generic constructor.

Vulkan.ViewportWScalingNV — Method

Generic constructor.

Vulkan.VulkanApplication — Type

A Vulkan application represents any program that uses the Vulkan API. Application goals and setups may vary greatly, but they all require many structures to function. Some may do offline rendering (without presenting the result to a window), or use Vulkan Compute and render directly to a display, but both uses share similarities in that they require an Instance and one (or several) devices.

Vulkan.XYColorEXT — Method

Generic constructor.

Vulkan.acquire_full_screen_exclusive_mode_ext — Method

Generic definition

Vulkan.acquire_next_image_2_khr — Method

Generic definition

Vulkan.acquire_next_image_khr — Method

Generic definition

Vulkan.acquire_performance_configuration_intel — Method

Generic definition

Vulkan.acquire_profiling_lock_khr — Method

Generic definition

Vulkan.acquire_xlib_display_ext — Method

Generic definition

Vulkan.allocate_command_buffers — Method

Generic definition

Vulkan.allocate_descriptor_sets — Method

Generic definition

Vulkan.allocate_memory — Method

Generic definition

Vulkan.begin_command_buffer — Method

Generic definition

Vulkan.bind_acceleration_structure_memory_khr — Method

Generic definition

Vulkan.bind_buffer_memory — Method

Generic definition

Vulkan.bind_buffer_memory_2 — Method

Generic definition

Vulkan.bind_image_memory — Method

Generic definition

Vulkan.bind_image_memory_2 — Method

Generic definition

Vulkan.build_acceleration_structure_khr — Method

Generic definition

Vulkan.cmd_begin_conditional_rendering_ext — Method

Generic definition

Vulkan.cmd_begin_debug_utils_label_ext — Method

Generic definition

Vulkan.cmd_begin_query — Method

Generic definition

Vulkan.cmd_begin_query_indexed_ext — Method

Generic definition

Vulkan.cmd_begin_render_pass — Method

Generic definition

Vulkan.cmd_begin_render_pass_2 — Method

Generic definition

Vulkan.cmd_begin_transform_feedback_ext — Method

Generic definition

Vulkan.cmd_bind_descriptor_sets — Method

Generic definition

Vulkan.cmd_bind_index_buffer — Method

Generic definition

Vulkan.cmd_bind_pipeline — Method

Generic definition

Vulkan.cmd_bind_pipeline_shader_group_nv — Method

Generic definition

Vulkan.cmd_bind_shading_rate_image_nv — Method

Generic definition

Vulkan.cmd_bind_transform_feedback_buffers_ext — Method

Generic definition

Vulkan.cmd_bind_vertex_buffers — Method

Generic definition

Vulkan.cmd_bind_vertex_buffers_2_ext — Method

Generic definition

Vulkan.cmd_blit_image — Method

Generic definition

Vulkan.cmd_build_acceleration_structure_indirect_khr — Method

Generic definition

Vulkan.cmd_build_acceleration_structure_khr — Method

Generic definition

Vulkan.cmd_build_acceleration_structure_nv — Method

Generic definition

Vulkan.cmd_clear_attachments — Method

Generic definition

Vulkan.cmd_clear_color_image — Method

Generic definition

Vulkan.cmd_clear_depth_stencil_image — Method

Generic definition

Vulkan.cmd_copy_acceleration_structure_khr — Method

Generic definition

Vulkan.cmd_copy_acceleration_structure_nv — Method

Generic definition

Vulkan.cmd_copy_acceleration_structure_to_memory_khr — Method

Generic definition

Vulkan.cmd_copy_buffer — Method

Generic definition

Vulkan.cmd_copy_buffer_to_image — Method

Generic definition

Vulkan.cmd_copy_image — Method

Generic definition

Vulkan.cmd_copy_image_to_buffer — Method

Generic definition

Vulkan.cmd_copy_memory_to_acceleration_structure_khr — Method

Generic definition

Vulkan.cmd_copy_query_pool_results — Method

Generic definition

Vulkan.cmd_debug_marker_begin_ext — Method

Generic definition

Vulkan.cmd_debug_marker_end_ext — Method

Generic definition

Vulkan.cmd_debug_marker_insert_ext — Method

Generic definition

Vulkan.cmd_dispatch — Method

Generic definition

Vulkan.cmd_dispatch_base — Method

Generic definition

Vulkan.cmd_dispatch_indirect — Method

Generic definition

Vulkan.cmd_draw — Method

Generic definition

Vulkan.cmd_draw_indexed — Method

Generic definition

Vulkan.cmd_draw_indexed_indirect — Method

Generic definition

Vulkan.cmd_draw_indexed_indirect_count — Method

Generic definition

Vulkan.cmd_draw_indirect — Method

Generic definition

Vulkan.cmd_draw_indirect_byte_count_ext — Method

Generic definition

Vulkan.cmd_draw_indirect_count — Method

Generic definition

Vulkan.cmd_draw_mesh_tasks_indirect_count_nv — Method

Generic definition

Vulkan.cmd_draw_mesh_tasks_indirect_nv — Method

Generic definition

Vulkan.cmd_draw_mesh_tasks_nv — Method

Generic definition

Vulkan.cmd_end_conditional_rendering_ext — Method

Generic definition

Vulkan.cmd_end_debug_utils_label_ext — Method

Generic definition

Vulkan.cmd_end_query — Method

Generic definition

Vulkan.cmd_end_query_indexed_ext — Method

Generic definition

Vulkan.cmd_end_render_pass — Method

Generic definition

Vulkan.cmd_end_render_pass_2 — Method

Generic definition

Vulkan.cmd_end_transform_feedback_ext — Method

Generic definition

Vulkan.cmd_execute_commands — Method

Generic definition

Vulkan.cmd_execute_generated_commands_nv — Method

Generic definition

Vulkan.cmd_fill_buffer — Method

Generic definition

Vulkan.cmd_insert_debug_utils_label_ext — Method

Generic definition

Vulkan.cmd_next_subpass — Method

Generic definition

Vulkan.cmd_next_subpass_2 — Method

Generic definition

Vulkan.cmd_pipeline_barrier — Method

Generic definition

Vulkan.cmd_preprocess_generated_commands_nv — Method

Generic definition

Vulkan.cmd_push_constants — Method

Generic definition

Vulkan.cmd_push_descriptor_set_khr — Method

Generic definition

Vulkan.cmd_push_descriptor_set_with_template_khr — Method

Generic definition

Vulkan.cmd_reset_event — Method

Generic definition

Vulkan.cmd_reset_query_pool — Method

Generic definition

Vulkan.cmd_resolve_image — Method

Generic definition

Vulkan.cmd_set_blend_constants — Method

Generic definition

Vulkan.cmd_set_checkpoint_nv — Method

Generic definition

Vulkan.cmd_set_coarse_sample_order_nv — Method

Generic definition

Vulkan.cmd_set_cull_mode_ext — Method

Generic definition

Vulkan.cmd_set_depth_bias — Method

Generic definition

Vulkan.cmd_set_depth_bounds — Method

Generic definition

Vulkan.cmd_set_depth_bounds_test_enable_ext — Method

Generic definition

Vulkan.cmd_set_depth_compare_op_ext — Method

Generic definition

Vulkan.cmd_set_depth_test_enable_ext — Method

Generic definition

Vulkan.cmd_set_depth_write_enable_ext — Method

Generic definition

Vulkan.cmd_set_device_mask — Method

Generic definition

Vulkan.cmd_set_discard_rectangle_ext — Method

Generic definition

Vulkan.cmd_set_event — Method

Generic definition

Vulkan.cmd_set_exclusive_scissor_nv — Method

Generic definition

Vulkan.cmd_set_front_face_ext — Method

Generic definition

Vulkan.cmd_set_line_stipple_ext — Method

Generic definition

Vulkan.cmd_set_line_width — Method

Generic definition

Vulkan.cmd_set_performance_marker_intel — Method

Generic definition

Vulkan.cmd_set_performance_override_intel — Method

Generic definition

Vulkan.cmd_set_performance_stream_marker_intel — Method

Generic definition

Vulkan.cmd_set_primitive_topology_ext — Method

Generic definition

Vulkan.cmd_set_sample_locations_ext — Method

Generic definition

Vulkan.cmd_set_scissor — Method

Generic definition

Vulkan.cmd_set_scissor_with_count_ext — Method

Generic definition

Vulkan.cmd_set_stencil_compare_mask — Method

Generic definition

Vulkan.cmd_set_stencil_op_ext — Method

Generic definition

Vulkan.cmd_set_stencil_reference — Method

Generic definition

Vulkan.cmd_set_stencil_test_enable_ext — Method

Generic definition

Vulkan.cmd_set_stencil_write_mask — Method

Generic definition

Vulkan.cmd_set_viewport — Method

Generic definition

Vulkan.cmd_set_viewport_shading_rate_palette_nv — Method

Generic definition

Vulkan.cmd_set_viewport_w_scaling_nv — Method

Generic definition

Vulkan.cmd_set_viewport_with_count_ext — Method

Generic definition

Vulkan.cmd_trace_rays_indirect_khr — Method

Generic definition

Vulkan.cmd_trace_rays_khr — Method

Generic definition

Vulkan.cmd_trace_rays_nv — Method

Generic definition

Vulkan.cmd_update_buffer — Method

Generic definition

Vulkan.cmd_wait_events — Method

Generic definition

Vulkan.cmd_write_acceleration_structures_properties_khr — Method

Generic definition

Vulkan.cmd_write_buffer_marker_amd — Method

Generic definition

Vulkan.cmd_write_timestamp — Method

Generic definition

Vulkan.compile_deferred_nv — Method

Generic definition

Vulkan.copy_acceleration_structure_khr — Method

Generic definition

Vulkan.copy_acceleration_structure_to_memory_khr — Method

Generic definition

Vulkan.copy_memory_to_acceleration_structure_khr — Method

Generic definition

Vulkan.debug_marker_set_object_name_ext — Method

Generic definition

Vulkan.debug_marker_set_object_tag_ext — Method

Generic definition

Vulkan.debug_report_message_ext — Method

Generic definition

Vulkan.deferred_operation_join_khr — Method

Generic definition

Vulkan.device_wait_idle — Method

Generic definition

Vulkan.display_power_control_ext — Method

Generic definition

Vulkan.end_command_buffer — Method

Generic definition

Vulkan.enumerate_instance_version — Method

Generic definition

Vulkan.flush_mapped_memory_ranges — Method

Generic definition

Vulkan.free_command_buffers — Method

Generic definition

Vulkan.free_descriptor_sets — Method

Generic definition

Vulkan.free_memory — Method

Generic definition

Vulkan.get_acceleration_structure_device_address_khr — Method

Generic definition

Vulkan.get_acceleration_structure_handle_nv — Method

Generic definition

Vulkan.get_acceleration_structure_memory_requirements_khr — Method

Generic definition

Vulkan.get_acceleration_structure_memory_requirements_nv — Method

Generic definition

Vulkan.get_android_hardware_buffer_properties_android — Method

Generic definition

Vulkan.get_buffer_device_address — Method

Generic definition

Vulkan.get_buffer_memory_requirements — Method

Generic definition

Vulkan.get_buffer_memory_requirements_2 — Method

Generic definition

Vulkan.get_buffer_opaque_capture_address — Method

Generic definition

Vulkan.get_calibrated_timestamps_ext — Method

Generic definition

Vulkan.get_deferred_operation_max_concurrency_khr — Method

Generic definition

Vulkan.get_deferred_operation_result_khr — Method

Generic definition

Vulkan.get_descriptor_set_layout_support — Method

Generic definition

Vulkan.get_device_acceleration_structure_compatibility_khr — Method

Generic definition

Vulkan.get_device_group_peer_memory_features — Method

Generic definition

Vulkan.get_device_group_present_capabilities_khr — Method

Generic definition

Vulkan.get_device_group_surface_present_modes_2_ext — Method

Generic definition

Vulkan.get_device_group_surface_present_modes_khr — Method

Generic definition

Vulkan.get_device_memory_commitment — Method

Generic definition

Vulkan.get_device_memory_opaque_capture_address — Method

Generic definition

Vulkan.get_device_proc_addr — Method

Generic definition

Vulkan.get_device_queue — Method

Generic definition

Vulkan.get_device_queue_2 — Method

Generic definition

Vulkan.get_display_plane_capabilities_2_khr — Method

Generic definition

Vulkan.get_display_plane_capabilities_khr — Method

Generic definition

Vulkan.get_event_status — Method

Generic definition

Vulkan.get_fence_fd_khr — Method

Generic definition

Vulkan.get_fence_status — Method

Generic definition

Vulkan.get_fence_win_32_handle_khr — Method

Generic definition

Vulkan.get_generated_commands_memory_requirements_nv — Method

Generic definition

Vulkan.get_image_drm_format_modifier_properties_ext — Method

Generic definition

Vulkan.get_image_memory_requirements — Method

Generic definition

Vulkan.get_image_memory_requirements_2 — Method

Generic definition

Vulkan.get_image_subresource_layout — Method

Generic definition

Vulkan.get_image_view_address_nvx — Method

Generic definition

Vulkan.get_image_view_handle_nvx — Method

Generic definition

Vulkan.get_instance_proc_addr — Method

Generic definition

Vulkan.get_memory_android_hardware_buffer_android — Method

Generic definition

Vulkan.get_memory_fd_khr — Method

Generic definition

Vulkan.get_memory_fd_properties_khr — Method

Generic definition

Vulkan.get_memory_host_pointer_properties_ext — Method

Generic definition

Vulkan.get_memory_win_32_handle_khr — Method

Generic definition

Vulkan.get_memory_win_32_handle_nv — Method

Generic definition

Vulkan.get_memory_win_32_handle_properties_khr — Method

Generic definition

Vulkan.get_performance_parameter_intel — Method

Generic definition

Vulkan.get_physical_device_external_buffer_properties — Method

Generic definition

Vulkan.get_physical_device_external_fence_properties — Method

Generic definition

Vulkan.get_physical_device_external_image_format_properties_nv — Method

Generic definition

Vulkan.get_physical_device_external_semaphore_properties — Method

Generic definition

Vulkan.get_physical_device_features — Method

Generic definition

Vulkan.get_physical_device_features_2 — Method

Generic definition

Vulkan.get_physical_device_format_properties — Method

Generic definition

Vulkan.get_physical_device_format_properties_2 — Method

Generic definition

Vulkan.get_physical_device_image_format_properties — Method

Generic definition

Vulkan.get_physical_device_image_format_properties_2 — Method

Generic definition

Vulkan.get_physical_device_memory_properties — Method

Generic definition

Vulkan.get_physical_device_memory_properties_2 — Method

Generic definition

Vulkan.get_physical_device_multisample_properties_ext — Method

Generic definition

Vulkan.get_physical_device_properties — Method

Generic definition

Vulkan.get_physical_device_properties_2 — Method

Generic definition

Vulkan.get_physical_device_queue_family_performance_query_passes_khr — Method

Generic definition

Vulkan.get_physical_device_surface_capabilities_2_ext — Method

Generic definition

Vulkan.get_physical_device_surface_capabilities_2_khr — Method

Generic definition

Vulkan.get_physical_device_surface_capabilities_khr — Method

Generic definition

Vulkan.get_physical_device_surface_support_khr — Method

Generic definition

Vulkan.get_physical_device_wayland_presentation_support_khr — Method

Generic definition

Vulkan.get_physical_device_win_32_presentation_support_khr — Method

Generic definition

Vulkan.get_physical_device_xcb_presentation_support_khr — Method

Generic definition

Vulkan.get_physical_device_xlib_presentation_support_khr — Method

Generic definition

Vulkan.get_private_data_ext — Method

Generic definition

Vulkan.get_query_pool_results — Method

Generic definition

Vulkan.get_rand_r_output_display_ext — Method

Generic definition

Vulkan.get_ray_tracing_capture_replay_shader_group_handles_khr — Method

Generic definition

Vulkan.get_ray_tracing_shader_group_handles_khr — Method

Generic definition

Vulkan.get_refresh_cycle_duration_google — Method

Generic definition

Vulkan.get_render_area_granularity — Method

Generic definition

Vulkan.get_semaphore_counter_value — Method

Generic definition

Vulkan.get_semaphore_fd_khr — Method

Generic definition

Vulkan.get_semaphore_win_32_handle_khr — Method

Generic definition

Vulkan.get_swapchain_counter_ext — Method

Generic definition

Vulkan.get_swapchain_status_khr — Method

Generic definition

Vulkan.import_fence_fd_khr — Method

Generic definition

Vulkan.import_fence_win_32_handle_khr — Method

Generic definition

Vulkan.import_semaphore_fd_khr — Method

Generic definition

Vulkan.import_semaphore_win_32_handle_khr — Method

Generic definition

Vulkan.initialize_performance_api_intel — Method

Generic definition

Vulkan.invalidate_mapped_memory_ranges — Method

Generic definition

Vulkan.map_memory — Method

Generic definition

Vulkan.merge_pipeline_caches — Method

Generic definition

Vulkan.merge_validation_caches_ext — Method

Generic definition

Vulkan.queue_begin_debug_utils_label_ext — Method

Generic definition

Vulkan.queue_bind_sparse — Method

Generic definition

Vulkan.queue_end_debug_utils_label_ext — Method

Generic definition

Vulkan.queue_insert_debug_utils_label_ext — Method

Generic definition

Vulkan.queue_present_khr — Method

Generic definition

Vulkan.queue_set_performance_configuration_intel — Method

Generic definition

Vulkan.queue_submit — Method

Generic definition

Vulkan.queue_wait_idle — Method

Generic definition

Vulkan.register_device_event_ext — Method

Generic definition

Vulkan.register_display_event_ext — Method

Generic definition

Vulkan.release_display_ext — Method

Generic definition

Vulkan.release_full_screen_exclusive_mode_ext — Method

Generic definition

Vulkan.release_performance_configuration_intel — Method

Generic definition

Vulkan.release_profiling_lock_khr — Method

Generic definition

Vulkan.reset_command_buffer — Method

Generic definition

Vulkan.reset_command_pool — Method

Generic definition

Vulkan.reset_descriptor_pool — Method

Generic definition

Vulkan.reset_event — Method

Generic definition

Vulkan.reset_fences — Method

Generic definition

Vulkan.reset_query_pool — Method

Generic definition

Vulkan.set_debug_utils_object_name_ext — Method

Generic definition

Vulkan.set_debug_utils_object_tag_ext — Method

Generic definition

Vulkan.set_event — Method

Generic definition

Vulkan.set_hdr_metadata_ext — Method

Generic definition

Vulkan.set_local_dimming_amd — Method

Generic definition

Vulkan.set_private_data_ext — Method

Generic definition

Vulkan.setup_viewport! — Method

Setup the application viewport.

Vulkan.setup_viewport! — Method

By default, the extent is taken as the application swapchain extent, but a custom value can be provided instead.

Vulkan.signal_semaphore — Method

Generic definition

Vulkan.submit_debug_utils_message_ext — Method

Generic definition

Vulkan.trim_command_pool — Method

Generic definition

Vulkan.uninitialize_performance_api_intel — Method

Generic definition

Vulkan.unmap_memory — Method

Generic definition

Vulkan.update_descriptor_set_with_template — Method

Generic definition

Vulkan.update_descriptor_sets — Method

Generic definition

Vulkan.wait_for_fences — Method

Generic definition

Vulkan.wait_semaphores — Method

Generic definition

Vulkan.write_acceleration_structures_properties_khr — Method

Generic definition

Vulkan.@check — Macro

@check vkFunctionSomething()

Checks whether the expression returned VK_SUCCESS or any non-error codes. Else, throw an error printing the corresponding code.