API

Vulkan.DebugUtilsMessengerEXT

Vulkan.Handle

Vulkan.ReturnedOnly

Vulkan.VulkanStruct

Vulkan.@check

Vulkan.DebugUtilsMessengerEXT — Method

Register a user callback and return the corresponding messenger.

A default named default_debug_callback can be converted to a function pointer to use as a callback.

Warning

callback must be a function pointer of type Ptr{Nothing} obtained from a callback_f function as follows: callback = @cfunction(callback_f, UInt32, (VkDebugUtilsMessageSeverityFlagBitsEXT, VkDebugUtilsMessageTypeFlagBitsEXT, Ptr{vk.VkDebugUtilsMessengerCallbackDataEXT}, Ptr{Cvoid})) with callback_f a Julia function with a signature matching the @cfunction call.

DebugUtilsMessengerEXT(instance::Instance, callback::Ptr{Nothing}; severity, types) -> DebugUtilsMessengerEXT

Vulkan.Handle — Type

Opaque handle referring to internal Vulkan data. Finalizer registration is taken care of by constructors.

abstract type Handle <: VulkanStruct{false}

Vulkan.ReturnedOnly — Type

Represents a structure that will never be requested by API functions.

abstract type ReturnedOnly <: VulkanStruct{false}

Vulkan.VulkanStruct — Type

Represents any kind of wrapper structure that was generated from a Vulkan structure. D is a Bool parameter indicating whether the structure has specific dependencies or not.

Vulkan.@check — Macro

@check vkFunctionSomething()

Checks whether the expression returned VK_SUCCESS or any non-error codes. Else, throw an error printing the corresponding code.