API

Vulkan.DebugUtilsMessengerEXT

Vulkan.Handle

Vulkan.PhysicalDeviceFeatures

Vulkan.ReturnedOnly

Vulkan.VulkanError

Vulkan.VulkanStruct

Vulkan.find_queue_index

Vulkan.@check

Vulkan.DebugUtilsMessengerEXT — Method

Register a user callback and return the corresponding messenger.

A default named default_debug_callback can be converted to a function pointer to use as a callback.

Warning

callback must be a function pointer of type Ptr{Nothing} obtained from a callback_f function as follows: callback = @cfunction(callback_f, UInt32, (VkDebugUtilsMessageSeverityFlagBitsEXT, VkDebugUtilsMessageTypeFlagBitsEXT, Ptr{vk.VkDebugUtilsMessengerCallbackDataEXT}, Ptr{Cvoid})) with callback_f a Julia function with a signature matching the @cfunction call.

DebugUtilsMessengerEXT(instance::Instance, callback::Ptr{Nothing}; severity, types) -> DebugUtilsMessengerEXT

Vulkan.Handle — Type

Opaque handle referring to internal Vulkan data. Finalizer registration is taken care of by constructors.

abstract type Handle <: VulkanStruct{false}

Vulkan.PhysicalDeviceFeatures — Method

Return a PhysicalDeviceFeatures object with the fields present in features set to true.

julia> PhysicalDeviceFeatures([])
PhysicalDeviceFeatures()

julia> PhysicalDeviceFeatures([:wideLines, :sparseBinding])
PhysicalDeviceFeatures(wideLines, sparseBinding)

PhysicalDeviceFeatures(features::AbstractArray) -> Any

Vulkan.ReturnedOnly — Type

Represents a structure that will never be requested by API functions.

abstract type ReturnedOnly <: VulkanStruct{false}

Vulkan.VulkanError — Type

Exception type indicating that an API function returned a non-success code.

struct VulkanError <: Exception

msg::AbstractString

return_code::Any

Vulkan.VulkanStruct — Type

Represents any kind of wrapper structure that was generated from a Vulkan structure. D is a Bool parameter indicating whether the structure has specific dependencies or not.

Vulkan.find_queue_index — Method

Find a queue index (starting at 0) from physical_device which matches the provided queue_capabilities. queue_capabilities can be combination of VkQueueFlagBits.

julia> find_queue_index(physical_device, VK_QUEUE_COMPUTE_BIT & VK_QUEUE_GRAPHICS_BIT)
0

find_queue_index(physical_device::PhysicalDevice, queue_capabilities::Any) -> Int64

Vulkan.@check — Macro

@check vkCreateInstance(args...)

If the expression does not return VK_SUCCESS, raise a VulkanError holding the return code.

Requires the ERROR_CHECKING preference enabled.